Changed parsing logic in request.py
One particular unit test showed that once a protocol or even a country code is being parsed
the protocol or country code will be succesfully added, but the previous logic would then
stop parsing. This is because the for loop will check for the next key word and will hit the
previously added tansport protocol next, obfs4 for example. But since it is not a key word it
would result into breaking out of the loop and stop the parsing process.
This fix added a mechanism that checks if a protocol or country code was found, count the index
and than skip the indexes that include the specific protocol or country code